updating column type in migration for ids

+++ Domain/Entities/DoorAccessLevel.cs
@@ -2,9 +2,9 @@
 {
     public class DoorAccessLevel
     {
-        public int DoorId { get; set; }
+        public long DoorId { get; set; }
         public Door Door { get; set; } = null!;
-        public int AccessLevelId { get; set; }
+        public long AccessLevelId { get; set; }
         public AccessLevel AccessLevel { get; set; } = null!;
     }
 }
