+++ Domain/Entities/UserAccessLevel.cs
@@ -2,9 +2,9 @@
 {
     public class UserAccessLevel
     {
-        public int AccessLevelId { get; set; }
+        public long AccessLevelId { get; set; }
         public AccessLevel AccessLevel { get; set; } = null!;
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
     }
 }
